[info] ajout cartes IO et Pneumatique au simu. COLLADALoader plus robuste, les transformations géométriques sont mieux gérées. Toumai est importé en entier. Les logs sont à nouveau en couleur.
Former-commit-id: 7e2a2b7ea56710f097d50a300a6fb4165a333b46

+++ commun/MathToolbox/Vector4.h
@@ -127,7 +127,7 @@ inline T dot(const Vector4<T>& v1, const Vector4<T>& v2) {
 // Affichage avec iostream
 template <class T>
 inline std::ostream& operator<<(std::ostream& os, const Vector4<T>& v) {
-	os << "[" << (int)v.x << ", " << (int)v.y << ", " << (int)v.z << ", " << (int)v.w << "]";
+	os << "[" << v.x << ", " << v.y << ", " << v.z << ", " << v.w << "]";
 	return os;
 }
